+++ crates/native/src/bridge_generated.rs
@@ -14,8 +14,10 @@
 
 use crate::api::*;
 use core::panic::UnwindSafe;
-use flutter_rust_bridge::{rust2dart::IntoIntoDart, *};
-use std::{ffi::c_void, sync::Arc};
+use flutter_rust_bridge::rust2dart::IntoIntoDart;
+use flutter_rust_bridge::*;
+use std::ffi::c_void;
+use std::sync::Arc;
 
 // Section: imports