+++ lib/src/platform/video_renderer.dart
@@ -18,12 +18,11 @@ class RTCVideoValue {
   });
 
   static const RTCVideoValue empty = RTCVideoValue();
+
   final double width;
   final double height;
   final int rotation;
   final bool renderVideo;
-
-  int get quarterTurnsRotation => (rotation / 90).round();
 
   double get aspectRatio {
     if (width == 0.0 || height == 0.0) {
@@ -61,8 +60,6 @@ abstract class VideoRenderer extends ValueNotifier<RTCVideoValue> {
   int get videoWidth;
 
   int get videoHeight;
-
-  int get quarterTurnsRotation;
 
   set mirror(bool mirror);