+++ lib/src/platform/web/video_renderer.dart
@@ -101,9 +101,6 @@ class WebVideoRenderer extends VideoRenderer {
   int get videoHeight => value.height.toInt();
 
   @override
-  int get quarterTurnsRotation => value.quarterTurnsRotation;
-
-  @override
   int get textureId => _textureId;
 
   @override