+++ lib/src/platform/web/video_view.dart
@@ -13,7 +13,6 @@ class VideoView extends StatefulWidget {
     this.mirror = false,
     this.enableContextMenu = true,
     this.filterQuality = FilterQuality.low,
-    this.autoRotate = false,
   }) : super(key: key);
 
   final VideoRenderer _renderer;
@@ -21,7 +20,8 @@ class VideoView extends StatefulWidget {
   final bool mirror;
   final bool enableContextMenu;
   final FilterQuality filterQuality;
-  final bool autoRotate;
+
+  WebVideoRenderer get videoRenderer => _renderer as WebVideoRenderer;
 
   @override
   State<VideoView> createState() => _VideoViewState();