+++ CHANGELOG.md
@@ -20,11 +20,17 @@ All user visible changes to this project will be documented in this file. This p
 - Refactor Audio Device Module to use [OpenAL] library for playout. ([#117])
 - Fire `onDeviceChange` callback whenever an output audio device is changed in system settings on desktop platforms. ([#119], [#120])
 - Upgraded [libwebrtc] to [116.0.5845.110] version. ([#123])
+- `VideoRenderer.width` and `VideoRenderer.height` now take rotation into account. ([#124])
+
+### Fixed
+
+- Video renderer stretching a picture after rotation. ([#124])
 
 [#117]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/117
 [#119]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/119
 [#120]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/120
 [#123]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/123
+[#124]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/124
 [#125]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/125
 [116.0.5845.110]: https://github.com/instrumentisto/libwebrtc-bin/releases/tag/116.0.5845.110